refactor: move utilities into separate directory
